requirements.txt
# PS-MCP 图片处理服务器依赖

# MCP协议支持
mcp>=1.0.0

# 图片处理核心库
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# 系统监控
psutil>=5.9.0

# 异步支持
asyncio-throttle>=1.0.0

# 可选：高级图片处理
scikit-image>=0.21.0
imageio>=2.31.0

# 开发和测试
pytest>=7.4.0
pytest-asyncio>=0.21.0